refactor(configuration): move utility functions and resolve logic to separate modules
Extract `is_config_path`, `is_path_to_existing_file`, `resolve_references`, and `find_unresolved_directives` from `config.py` into dedicated modules to improve code organization and maintainability. Update imports across the codebase to reference the new module locations.
